url_test/pyproject.toml
[project]
name = "html-viewer"
version = "0.1.0"
description = "A Streamlit app to fetch and display prettified HTML from a URL."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "requests",
    "beautifulsoup4",
    "trafilatura (>=2.0.0,<3.0.0)",
    "newspaper3k (>=0.2.8,<0.3.0)",
]

[project.scripts]
html-viewer = "app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
